The telescope subsystem stuff is fr made. It should to the yeah and what not like elevator but it's a telescope. Credits: -Josh -Charles
