+++ src/main/java/frc/robot/subsystems/telescope/TelescopeIOSparks.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.TelescopeIO;
+package frc.robot.subsystems.telescope;
 
 import com.revrobotics.CANSparkMax;
 
@@ -7,10 +7,10 @@ public class TelescopeIOSparks implements TelescopeIO{
     private CANSparkMax left, right;
 
     @Override
-    public void updateInputs(TelescopeInputs inputs) {        
+    public void updateInputs(TelescopeInputs inputs) {
     }
 
     @Override
-    public void setVoltage(double volts) {        
+    public void setVoltage(double volts) {
     }
 }
